requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifultable==1.1.0
certifi==2024.6.2
click==8.1.7
dict2xml==1.7.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orderedmultidict==1.0.1
orjson==3.10.3
pgMigrationMgr @ git+https://github.com/Lucas-Palomo/pgMigrationMgr@62a956765523fb0907d199e82bc9c28ec3af516d
prompt_toolkit==3.0.46
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyPika==0.48.9
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
simplejsondb==0.4.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
